fix: project max_parallel_tasks is omitted by the API when empty. (#49)
Fixes: #30 

The Semaphore project began omitting `max_parallel_tasks` in an the API
response
([here](https://github.com/semaphoreui/semaphore/commit/61ef397084ffe563fa54f98e0afe3c3872e38a9e)),
but the provider expected it to be present. This change wasn't
documented in the Semaphore API spec. A copy of the API was created in
this repo and the necessary changes added. This should fix the issues
with Semaphore 2.13 and 2.14.